requirements.txt
opencv-python
pillow
numpy
flask
ultralytics
PyTorch
stockfish
